Partially fix neighbor validation
Neighbor is only updated if it either hasn't been visited yet, or:

1. has a new cost-to-reach that is lower than its previous registered
    one, and
2. Has a new cost-to-reach + cost-to-go that is less than the maximum
   trajectory length
Previously, the cost-to-go was taken from the Map() whether or not it
was even found; the value is initialized in advance just in case.

Found crazy bug! With a maximum range of 51.7, the far left point can be
reached, but with a value of 52 or 60, it can't; the max_range value
seems to be messing with how A-star functions, but I'm not quite sure
how yet.
Side-note, adding the non-penalized cost-to-reach to the to_visit queue
avoids this issue, so the penalized cost-to-reach and max range are
together causing some behavior to go haywire.