Add backup planner in case of failure
A-Star seems to be incompatible with both a high risk
multiplier/penalty and a maximum range; under particular circumstances,
it can fail to find a path when one should be found, namely when the
maximum range is very close to the direct path between the nest and
site.
Since this type of trajectory necessarily requires mostly ignoring the
high risk penalty, a backup path planner was added which behaves exactly
the same, but without the high risk penalty.
I *think* this works under all circumstances (unless, of course, the max
range is impossibly small regardless of risk zones).